Verify all 80 DSA entries, fix reference bugs, and add explained solutions
Add independent behavioral tests for 240 C++ implementations, sanitizer checks,
problem walkthroughs, platform contract notes, and continuous verification.

+++ 01_Linear_Traversal_and_In_Place/GeeksforGeeks/GFG_Second_Largest/04_optimal_solution.cpp
@@ -4,8 +4,8 @@ using namespace std;
 class Solution {
 public:
     int getSecondLargest(vector<int> &arr) {
-        int largestValue = INT_MIN;
-        int secondLargest = INT_MIN;
+        long long largestValue = LLONG_MIN;
+        long long secondLargest = LLONG_MIN;
         for(int i = 0; i < arr.size(); i++) {
             if(arr[i] > largestValue) {
                 secondLargest = largestValue;
@@ -15,7 +15,7 @@ public:
                 secondLargest = arr[i];
             }
         }
-        if(secondLargest == INT_MIN) {
+        if(secondLargest == LLONG_MIN) {
             return -1;
         }
         return secondLargest;
